Decluttered the code in constructQFT(N) by removing verbose mode.
Also changed the inner loop in constructQFT(N) to iterate j from i down to 1 instead of from 0 up to i - 1 to match the inner loop in the quantumFourierTransform() function in qc.py.

Added an explanation of how the constructQFT(N) algorithm works.

Removed constructQFT3() which I was using to test the algorithm that's now generalized by constructQFT(N).